+++ R/aggregate.R
@@ -30,7 +30,7 @@ aggregate.sf = function(x, by, FUN, ..., union = FALSE) {
 #' @param extensive logical; if TRUE, the attribute variables are assumed to be spatially extensive (like population) and the sum is preserved, otherwise, spatially intensive (like population density) and the mean is preserved.
 #' @examples
 #' nc = st_read(system.file("shape/nc.shp", package="sf"))
-#' g = sf:::st_makegrid(nc, n = c(20,10))
+#' g = sf:::st_make_grid(nc, n = c(20,10))
 #' a1 = st_interpolate_aw(nc["BIR74"], g, extensive = FALSE)
 #' sum(a1$BIR74) / sum(nc$BIR74) # not close to one: property is assumed spatially intensive
 #' a2 = st_interpolate_aw(nc["BIR74"], g, extensive = TRUE)