+++ tests/geos.R
@@ -35,7 +35,7 @@ st_combine(nc)
 st_dimension(st_sfc(st_point(0:1), st_linestring(rbind(c(0,0),c(1,1))), 
 	st_polygon(list(rbind(c(0,0), c(1,0), c(1,1), c(0,1), c(0,0))))))
 
-g = st_makegrid(nc)
+g = st_make_grid(nc)
 x = st_intersection(nc, g)
 x = st_intersection(g, nc)
 
@@ -46,7 +46,7 @@ set.seed(13531) # make reproducible
 
 st_line_sample(ls, density = 1, type = "random")
 
-g = st_makegrid(nc, n = c(20,10))
+g = st_make_grid(nc, n = c(20,10))
 
 a1 = st_interpolate_aw(nc["BIR74"], g, FALSE)
 sum(a1$BIR74) / sum(nc$BIR74) # not close to one: property is assumed spatially intensive
@@ -54,10 +54,10 @@ a2 = st_interpolate_aw(nc["BIR74"], g, extensive = TRUE)
 sum(a2$BIR74) / sum(nc$BIR74)
 
 # missing x:
-g = st_makegrid(offset = c(0,0), cellsize = c(1,1), n = c(10,10))
+g = st_make_grid(offset = c(0,0), cellsize = c(1,1), n = c(10,10))
 
 mls = st_multilinestring(list(rbind(c(0,0), c(1,1)), rbind(c(2,0), c(1,1))))
-st_linemerge(mls)
+st_line_merge(mls)
 
 if (sf_extSoftVersion()["GEOS"] >= "3.5.0") {
  # voronoi: